backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.31.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[argon2,bcrypt]==1.7.4
email-validator==2.2.0
orjson==3.10.7
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
requests==2.32.3
